Getting rails-observers temporarily from private branch.
I need to do this for now because rails-observers in incompatible
with the 4.0 release of the Rails stack (e.g. ActiveRecord, ActiveSupport)
because it's pegged to the beta. This is causing problems in the Cequel
gem which we are trying to upgrade to 4.0, but needs consistent versions
of these dependencies. This hack is temporary until the rails 4.0
upgrade pull request is merged in the source repo.